Update LICENSE to AGPL-3.0
This pull request updates the license of this project from GNU General Public License v3.0 (GPL-3.0) to GNU Affero General Public License v3.0 (AGPL-3.0).

We at Ultralytics have decided to make this change in order to better protect our intellectual property and ensure that any larger work that incorporates YOLO and is provided as a service over a network will also be open-sourced under AGPL-3.0.

AGPL-3.0 is very similar to GPL-3.0, but with an additional clause to address the use of software over a network. This change ensures that if someone uses YOLO as part of a larger work and provides that larger work as a service over a network (e.g., through a web application or API), they must also make the source code of the larger work available to users of the service under the AGPL-3.0 license.

We believe that this change will promote a more collaborative environment and help drive further innovation within the YOLO community.